refactor: shared layer로 lib 및 type 파일 이동
